manager-livestream/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
obsws-python>=1.7.0
Pillow>=10.0.0
pytesseract>=0.3.10
